TINYMCE-14493: split tests into separate file
