+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/sizing/ResizeTest.ts
@@ -1,6 +1,6 @@
 import { FocusTools, Keys, UiFinder, Waiter } from '@ephox/agar';
-import { before, beforeEach, context, describe, it } from '@ephox/bedrock-client';
-import { Attribute, Css, Focus, SugarBody, SugarElement } from '@ephox/sugar';
+import { before, beforeEach, describe, it } from '@ephox/bedrock-client';
+import { Css, SugarBody, SugarElement } from '@ephox/sugar';
 import { TinyDom, TinyHooks, TinyUiActions } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
@@ -124,54 +124,5 @@ describe('browser.tinymce.themes.silver.editor.sizing.ResizeTTest', () => {
       `Editor's height: 300 pixels, Editor's width: 300 pixels`,
       'aria-valuetext should have updated editors dimentions after resize'
     );
-  });
-
-  context('TINYMCE-14493: Vertical resize exposes aria-valuenow', () => {
-    const verticalHook = TinyHooks.bddSetup<Editor>({
-      base_url: '/project/tinymce/js/tinymce',
-      resize: true,
-      min_height: 300,
-      max_height: 500,
-      height: 400,
-      width: 400
-    }, []);
-
-    it('TINYMCE-14493: aria-valuenow updates when resizing vertically with the keyboard', async () => {
-      const editor = verticalHook.editor();
-      const container = TinyDom.container(editor);
-      await UiFinder.pWaitForVisible('Wait for resize handle to be visible', container, '.tox-statusbar__resize-handle');
-      const resizeHandle = UiFinder.findIn<HTMLElement>(container, '.tox-statusbar__resize-handle').getOrDie();
-      Focus.focus(resizeHandle);
-
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), '400', 'aria-valuenow should be set to editors height initially');
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), `${container.dom.offsetHeight}`, 'aria-valuenow should equal the container height');
-
-      for (let i = 0; i < 3; i++) {
-        TinyUiActions.keystroke(editor, Keys.down());
-      }
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), '460', 'aria-valuenow should increase by 3 * 20px after three down presses');
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), `${container.dom.offsetHeight}`, 'aria-valuenow should equal the container height');
-
-      for (let i = 0; i < 3; ++i) {
-        TinyUiActions.keystroke(editor, Keys.up());
-      }
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), '400', 'aria-valuenow should return to the original value after three up presses');
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), `${container.dom.offsetHeight}`, 'aria-valuenow should equal the container height');
-    });
-
-    it('TINYMCE-14493: aria-valuenow updates when resizing vertically by dragging', async () => {
-      const editor = verticalHook.editor();
-      const container = TinyDom.container(editor);
-      await UiFinder.pWaitForVisible('Wait for resize handle to be visible', container, '.tox-statusbar__resize-handle');
-      const resizeHandle = UiFinder.findIn<HTMLElement>(container, '.tox-statusbar__resize-handle').getOrDie();
-
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), '400', 'aria-valuenow should be set to editors height initially');
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), `${container.dom.offsetHeight}`, 'aria-valuenow should equal the container height');
-
-      await resizeEditorBy([ 0, 50 ], editor);
-
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), '450', 'aria-valuenow should reflect the new container height after dragging');
-      assert.equal(Attribute.get(resizeHandle, 'aria-valuenow'), `${container.dom.offsetHeight}`, 'aria-valuenow should equal the container height');
-    });
   });
 });
